+++ core/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `assert_as_serialize_arg_with()`, `assert_as_serialize_arg_eq()` and
   `assert_as_serialize_arg_eq_with()`, which can be used to assert that the `AsSerializeArg` and
   `Deserialize` impls on a type or pair of types match correctly.
+- Implement `serde::Serialize` and `Deserialize` for `Bytes`.
 
 ### Fixed